Downgrade error on only finding other architectures of a binary module dependency to a warning
Previously, for some module dependency 'Foo', which is a binary module with target-specific variants in the top-level 'Foo.swiftmodule' directory, if the current compilation's target is not one of the variants, the compiler (module loader) would emit an error.

For example when building client source targeting arm64-apple-macos:
'''
import Foo
'''
Where
'''
Foo.swiftmodule
  > x86_64-apple-macos.swiftmodule
  > i386-apple-macos.swiftmodule
'''
Would result in a diagnostic:
'''
error: could not find module 'Foo' for target 'arm64'; found: 'x86_64', 'i386', at: ...
'''
Which would halt module lookup and compilation. Crucially, this error diagnostic also got emitted when performing lookup for a 'canImport(Foo)' statement, which does not make sense since a failure to locate a module on a 'canImport' conditional should proceed silently by evaluating this conditional to 'false'.

This change switches this diagnostic to a warning, such a configuration likely represents an error in dependency layout and configuration, and the missing dependency module error diagnostic is still guaranteed to be emitted downstream.

Resolves rdar://161175498